# Optimize quantized conv2d for arm