<!--- 👉👉👉 Provide a general summary of your changes. 👈👈👈 -->

| |DE10-Nano (CPU)|Raspberry Pi 3B|
|----|----|----|
|Before|61.0ms|19.0ms|
|After|59.3ms|17.8ms|

Raspberry Pi result much faster than previous results ( #390 ), it comes from rebooting my Raspberry Pi... 

## Motivation and Context
<!--- 💭💡💭 Why is this change required? What problem does it solve? 💭💡💭 -->
<!--- 🔗 If it fixes an open issue, please link to the issue here. 🔗 -->

## Description
<!--- 📝🎯📝 Describe your changes in detail. 📝🎯📝 -->

## How has this been tested?
<!--- 🙆👌🙆 Please describe in detail how you tested your changes. 🙆👌🙆 -->
<!--- Include details of your testing environment, details of your manual tests, snippet code or commands of your manual tests, table of experiments results metrics, tests ran to see how your change affects other areas of the code, etc. -->
<!--- If you have experiments report documents please link to here. -->

```shell
$ PYTHONPATH=python/dlk python python/dlk/scripts/generate_project.py -i minimal_graph_with_shape.pb -o tmp/ -p object_detection -hq -ts -cache
$ cd tmp/object_detection.prj
$ mkdir build_arm
$ cd build_arm
$ cmake -DTOOLCHAIN_NAME=linux_arm -DUSE_NEON=1 ..
$ make -j8 lm
$ cd ..
$ mkdir build_aarch64
$ cd build_aarch64
$ cmake -DTOOLCHAIN_NAME=linux_aarch64 -DUSE_NEON=1 ..
$ make -j8 lm
```
and run executables.

## Screenshots (if appropriate):

## Types of changes
<!--- 🏁🎌🏁 What types of changes does your code introduce? Put an `x` in all the boxes that apply: 🏁🎌🏁 -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature / Optimization (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)

## Checklist:
<!--- ✅✅✅ Go over all the following points, and put an `x` in all the boxes that apply. ✅✅✅ -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->

<!--- TODO(wakisaka): After decide our code style, add this.
- [ ] My code follows the code style of this project.
-->
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.