@iizukak 2019-11-18
@tkng Can you review this PR?
I tried to review but it's difficult...